This the Final Project of the IST 4320 Class
1. Include what was the purpose of your app, what you attempted, what worked and what didn't work, what you would still like to do in the future. Also what did you learn?

The purpose of this app is to make a Tic Tac Toe app that is able to score the win/lose tally, with an opponent that is albeit basic since it utilizes the randomizer, it still works, there are also some stylistic additions done to make it look nice such as font changes, background and foreground colors changed to suit a dark mode-like aesthetic. I tried to unify the two conditions that determine whether they win or not with little success, so I kept it as two separate definitions that check on both the opponent's side and the player's side.


2. Any other feedback that you think would be helpful for me to understand your project or improve the project in the future?

There are definitely some improvements that the project can go through such as an actual ai that isn't just a random number generator, a separate button to add a light mode/dark mode and even some difficulty settings that determine whether the opponent will actually try or whether it'll go first, and an option for a bigger tic tac toe board that uses more than 9 panels for longer gameplay.